move acis api logic to acis.js
